Add `partition` and `interface` keywords
Add new keywords `partition` and `interface` to nv syntax.
`partition` takes a node and returns an `int8`.
`interface` takes an edge and returns an `option<aty>`.

partition and interface have also been added as fields to the
Syntax.network type, but not yet for the Syntax.srp_unfold type.

Currently, the transformations on the network simply erase or ignore
partition and interface information: handling these is TODO.