Add impls for init, trans, merge transformations
Add finished implementations for updating init trans and merge.
One possible improvement would be to give the hypothesis instead of the
init value for inputs in init, but trans and merge are otherwise done.

Add an additional field `broken` to OpenAdjGraph so we can keep track of
which edges were cut by partitioning. This is also used in
transform_trans to imitate the old edge that was used as part of the
transfer function.