Refactor OpenAdjGraph and onetwork types
This commit is to simplify some of the types used for partitioning.
Previously, we had two kinds of network: the original Syntax.network and
the partitioned form, Partition.onetwork.
Essentially, onetwork provides the same fields as network, but used an
OpenAdjGraph.t instead of an AdjGraph.t, which itself is simply an
AdjGraph.t with some extra fields.

This seemed too complicated to me and like it would cause annoyances
when trying to update the SMT and simulation tools to understand
partitioned graphs.
To solve the problem, I broke the extra fields of OpenAdjGraph.t out
into their own type (currently called "interfaces", but pending
bikeshedding).
The idea is that we can then have an onetwork type that simply contains
the original Syntax.network and the interfaces type.
This required refactoring some of OpenAdjGraph, but should give us
better mileage moving forward.