Add initial code to transform network
Add code to Main_defs to transform network when -k option given.
Put in remaining code to transform assert function in Partition.open_network.

I've commented out is_argty_func since it doesn't seem to match things
as expected. Printing the result for a given transfer function gave the
value I expected, but perhaps the information is not being saved in the
way I had anticipated, or something else is going on. In any case, the
function does nothing now so may be worth removing.