Reverse order of UnboxEdges and Kirigami
Move Kirigami code before UnBoxEdges code so that the edge unboxing
transform doesn't mess with Kirigami's expectation that the functions
are still talking about edges.

Generally, the hope is that the Kirigami transform can happen at the
earliest part of the pipeline, so some work needs to be done to ensure
that this is the case.
Currently, the code breaks within a `transform_exp` operation on a
function, presumably one of the ones introduced by Kirigami.
More work is needed to identify which function(s) this could be.