Fix require predicates and annotate types
Change partition_interface code to get the wrapped value of the VOption
directly to add to the interface, rather than wrapping the option.
This fixes a bug in the require clauses being generated, which referred
to "Some x" instead of "x".

Add type annotations in several places to help overcome the current
issues running nv with kirigami on.
Essentially what's happening is that the file is read in, parsed and
evaluated, and then the kirigami updates are made, which involves adding
new functions which haven't gone through the various transformations
that happened earlier.
This breaks an invariant of the final UnboxEdges transform, which
expects the various expressions to be typed already when it runs its
transformations, leading to many errors caused by unwrapping Nones.

There are two solutions to this problem:

1. Add type annotations everywhere in the kirigami code to ensure that
these invariants are respected when Transformers receive the network.
A possible advantage of this approach is that we know that the network
has been properly constructed and if bugs appear we can identify
kirigami as the culprit.
The downside is that this approach means we miss out on all the
transformations that were run earlier, which can cause problems
depending on the types of values used (e.g. if things get unboxed or
unwrapped).

2. Move the kirigami code to happen immediately after the file is parsed
into a list of declarations. In this scenario, we directly update the
declarations as those the file read specified different information.
This can potentially do weirder things if the kirigami code does
something bad, but allows us to take full advantage of the
transformations that are run later.

It looks at this point like moving to a solution 2 approach would be
easiest in the long term.