Implement partitioning at declaration stage
Implement solution 2 approach to handling partitioning in NV.
Currently fails due to typing errors.

Essentially the work has moved to a new open_declarations function,
which works the same way as open_network but operates over the list of
declarations rather than the network.
As a result, it doesn't have access to a true graph structure and does
some ugly enumeration to generate the additional nodes & edges it needs
for the partitioned graph.
Everything else works essentially the same way as before.