Make partition creation work hackily
We are now transforming declarations after the wellformedness check and
performing type inference a second time to handle all the cases that
were not done in TransformDecls.
This would be better to do carefully in the code, but for now is to-do.

Fix the input hypothesis map creation, which was given the base-base
edges before which didn't conform to transform_init's expectations that
it would get the input-base edges.
This solves a bug in setting the initial values of the input nodes.

Merge currently needs to be fixed. A selective merge causes problems
when a node u has some initial value a, and it has an output u_out whose
initial value is also a. If u's transfer returns a value b s.t. a < b,
then u_out will not update its value according to the merge function.
Changing merge to instead use the value of u directly at u_out sets
u_out to a instead of the solution at u, due to how the SMT constraints
work.
More work is needed to determine whether there is a formulation of merge
that will work generally or whether we need to also change the init for
output nodes.