Update transforms to fit expected types
Change the transfer declaration transformer to create a binary function
instead of a unary function.
This fixes an error in how the encoding into SMT works.

Change the merge declaration transformer to not create a group of
branches for the input nodes.
This was redundant because the input nodes never have in-edges, so merge
is never used.

Note that the merge case for the output branches is relying on the
particular implementation we use of merge, where the first variable is
the value at the node and the second is the value received.
In general, this is not a safe assumption to make due to #40:
merge is supposed to be commutative, so `merge x y = merge y x`.
A better solution to handling this (which we can investigate in the
future) is to add some information to the attribute to track the case of
whether or not the attribute is being sent along a (base,output) edge:
in those cases, we want to always select that particular attribute.
Nonetheless, it's not likely that this will break soon since output
nodes only have a single in-edge, so the merge will always just rely on
that one piece of information, and in some cases the original merge
function has the same effect here (e.g. base value is Some x, output
value is None, use mapbetter).