Allow predicate interfaces in kirigami
Allow users to give predicates instead of values for the interface.
The type must be TOption (TArrow attr_type TBool) instead of
TOption attr_type.
We now add the requires towards the end of the file instead, so that the
interface can reference other functions defined earlier.
(This would become unnecessary once the declarations are topsorted in #39)