Add an example partitioning with multiple solns
Add an example partition case where there are 2 solutions to the network
that both need to be in the hypotheses in order for the i-checks to
pass.

Also noticed a weird bug while creating this example with the handling
of partition.
Writing a partition function as follows:

    let partition node = match node with
      | 0n -> 0u8
      | 1n -> 1u8
      | 2n -> 1u8

causes the partition code to also split the 1-2 edge, even though they
have the same values.
This then leads to a typing error later in the code.
Using a wildcard to match the last two cases resolves this issue,
but it's not immediately clear to me why.