Attempt to fix travis errors
Reorganize the usage of dune and the naming scheme to try and stop
getting errors on Travis.
The cause of the errors we were seeing is that Travis wants to pull in
Nv_lib for the tests, but Nv_lib is not a dependency of nv.opam (it's a
part of it), so the opam install fails.
To get around this problem, I went back through the NV files and copied
the design used in the dune github examples for their hello_world
example.
I made everything use the same name of nv (rather than having nv_lib
exist) and removed the old dune-project and nv_lib.opam from src/lib.
This appears to work on my Linux machine when I do opam pin add nv .
and opam install nv -t -v.
Let's hope it works!