Add commands to divide nodes and edges across SRPs
Add commands to OpenAdjGraph to divide a list of nodes or edges across
multiple SRPs, enabling us to create multiple SRP instances with the
different nodes.
Currently this does not do anything for the cross-partition edge case;
that can be handled later or we can add more code to generate new edges
and nodes for cross-partition edges.

Other fixes:
- fix dune for kirigami (used nv_lib, changed to nv)
- remove unused function to check function types in TransformDecls, as
it wasn't working before and was mostly to just warn us early if there
were errors