Update main to run on a list of files
Reorganize parse_input to return a list of input tuples
which are handled by Main.ml in sequence.
This shouldn't change any of the original examples,
only the fancy partition ones (for now, although passing multiple files
could be a nice feature at some point).

Replace interfaces with interfaces_alt, a new struct
that can be constructed by divide_edges.
The only real difference is that the broken field has
been split into two maps, where outs_broken points to the broken edge
that the out node refers to, while broken_ins points to the in node
that the broken edge refers to: you can see that you could combine the
two to get back the original functionality.

Doing this has (aptly) broken a lot of existing Partition.ml code.
I've gone and tried to fix the uses of interfaces_alt in TransformDecls,
but my guess is that they are not quite correct because we aren't
properly handling the way that the new nodes and edges change some of
the helper maps like edge_hyps or input_exps yet.
This can also be seen from the fact that the code does not currently run
because an exception is thrown when trying to access the edge_hyps map
with an edge that doesn't exist.
Hence, next steps will involve correcting these issues.