More work to try and fix the helper maps
Some more work has been done to try and fix the maps that are used to
help transform the various functions.
Some screwy stuff is happening now that the outputs and inputs need a
different set of predicates for a different set of broken edges from the
interface function.

Currently, there is a type unification error when running
partition-fattree-10.nv.