Fix up SrpRemapping; bring into Partition.ml
Still seeing some issues with how the partitioned_srp is being used to
transform the SRPs.
We can see in the partition-simple example that we aren't properly
constraining the hypotheses just yet.
The next step is to run a toplevel transformer over the whole
declarations list to remap the nodes in an organized way, rather than
relying on complicated functions that call saved versions of the
original functions.
This will fix problems with internal functions not being called
correctly once things have been renamed.