Add prototype transformer impl for SRP remapping
The prototype implementation just handles remapping node and edge
patterns and values.
It is currently untested and may cause all kinds of weird issues.
One main case that is not properly handled at present is that, while we
can remap nodes/edges that do have a remapping, some nodes/edges don't.
These cases may need to be dropped from the graph entirely so that bugs
don't occur (or maybe they won't do anything?).

We also don't have a mapback transformer implemented yet.
This may not be desirable since the SRP is being split, although it
would help a user debug the splits.
To do this well, we'd need to be able to disambiguate which partition a
node falls into: should each new node also keep track of which SRP it's
in, so we can unambiguously tell them apart?
This is getting increasingly complicated.