Rewrite transform* fns to remap base nodes
While transforming using the transformers module would be perhaps the
nicest way to include the node remapping, it requires some fairly
sophisticated changes, like deleting branches that don't refer to any
base node in the new graph.

As an alternative, we can simply remap every node and edge on entry to
the transformed functions, so that everything internally looks as it did
in the original graph.
One downside of this is that we can't check for wellformedness of the
network anymore.

This new code gets us most of the way to this remapping, but it still
has some small issues where certain cases aren't being added when we
remap.