Fix transfer function not using new edges
transform_trans was still using the old edges in wildcard case;
update changes it to now properly remap the edges.

We are still seeing a lot of slowness in the preprocessing stage, and
also some possible logic errors when handling sp4.nv (possibly
insufficient interface annotations).
Sensible next steps would be:
0. Change partitioning to assume the input has been preprocessed, or at
least not to break if the input has been preprocessed.
1. Perform partitioning after preprocessing.
2. Extract remapping out to separate functions in the style of toEdge.