Update not to add output nodes
Update the transformer code to not add output nodes.
Instead, it sets the predicate given to also be over the base node
that was originally connected to the output node.
This causes a minor problem however, as we need to apply the predicate
to the applied transfer function of the original edge.
Rather than doing `n -> (pred x) && (old_assert n x)`,
we instead want to do `n -> (pred (old_trans e x)) && (old_assert n x)`,
which requires slightly more tooling.

My hope is that we can interpret the old transfer to simplify somewhat,
but it's still not ideal: the main benefit is that this halves the
number of nodes we need to create in order to get the network to be
verifiable partitioned, and it also obviates having to create a bad
merge case for that node.

In order to be able to continue developing this code, it may also be
time to pull in the solve branch, which gives some new features we can
use to make it easier to refer to the init, trans and merge functions
without having to perform these complicated steps to refer to them.