Explore issues with type inference after partition
Type inference is causing a problem after partitioning and partially
evaluating expressions, owing to the fact that partial evaluation can
lead to some expressions that don't have obvious type information.
An example of this case: after partially evaluating a branch of the init
method in fat4pol.nv, we get code that states:
```
let b~1961 = Some(...) in
  let fib~1962 = match (None, Some(1), None, b~1961) with
   | (None, None, None, None) -> None
   | _ -> ... match b~1961 with
         | None -> None
         | Some b -> Some Tuple5Get1-1 b in ...
```

Because we ignore the values of some of the options in fib~1962, this
ends up leaving the types of these options as unbound, which then causes
the mapUnrolling transformation not to know what to transform the types
to.

One strategy to try and fix this would be to properly go through and
relabel the types in TransformDecls, so that we don't need to call type
inference again to fix them.
Other options may exist to be explored.