Refactor partitioned_srp for simplicity
Refactor partitioned_srp data structure to remove a lot of extra code
that was needed to manipulate all the maps, allowing us to instead only
ever need to use one map at a time (I think).
Also refactor the transform code to follow a standard format, and add
attribute type as an argument so that we can make sure the type-checking
can be done properly so that no new type inference is needed afterwards.
This last aspect is not yet finished.