Update kirigami to use the new NV syntax
Update kirigami to use solution and assert syntax. Assertions as
described will cause some problems for current examples, however, as we
can no longer rely on the transformer to help us avoid indexing a node
that we don't want to test.
For instance, if a network has two nodes 0 and 1, and we cut it into two
partitions, we will now have two networks with nodes 0-2 each.
Any assertions made about the other node (0 or 1) will be applied to an
input or output node instead.
There isn't any way around this until we implement some rewriting
functionality that allows us to skip checking certain nodes or allows
foldNodes not to inspect input and output nodes.