Begin a nodeless partition impl
Using explicit nodes and edges to represent the input and output nodes
and edges leads to some problems when dealing with code that is
intended to run over all nodes/edges (e.g. foldNodes, foldEdges) but is
now being supplied nodes and edges the user did not intend.

The simple solutions of input and output nodes suggest, however, that we
could "inline" all of their behaviors to the relevant base nodes: the
init of a base node would be updated to capture its inputs, and an
assertion would be added to capture its outputs. Doing this sidesteps
any issues with using foldNodes or foldEdges.

This commit begins to handle that process but leaves it unfinished (and
hence this code will not work for verifying partitioned networks). What
remains to do is determine a way to handle a complication on the input
side: since there's no limit on the number of inputs a node may have, it
may have a very large number of symbolic values it needs to merge in to
its input, which requires some degree of care as the merge function is
binary.
One possible means to address this may be to add a recursive function
that, given a list of symbolic variables and a "calling node",
successively computes a new expression crafted from inlining a function
merging all these variables together.
Another possible way to address this could be through adding additional
require clauses for composite symbolic variables: if node 1 receives
hyp_21 and hyp_31, we can produce a new symbolic variable hyp_21_31
which has a single require clause that restricts it to the merge of
hyp_21 and hyp_31.