Start introducing Kirigami solver
Add code to change SMT encoding for Kirigami to directly split up
assertions and requires as necessary, depending on the checks we want to
perform.
The approach involves sorting the assertions and requires into 2 groups:
ranked hypotheses and guarantees get checked together as part of the
first encoding, and then hypotheses and properties get checked as part
of the second encoding.