Fix transcomp support within Partition
Change transcomp to no longer be a field of the partitioned_srp, and
properly introduce support for it on both init and output assertions,
where both transform_init and outputs_assert take a Syntax.exp, which
has the type of a transfer function, and use it to perform the transfer.
Supply an identity function for the alternative case.

Next step would still be to lift it out to being written in the file.