Change divide_decls to return a declarations tuple
Partition.divide_decls now returns a tuple, where the first element is
the transformed declarations from the original network, and the second
element is the added declarations for transforming the solution using
Kirigami, where we add new require and assertion clauses.
Symbolics are currently being added to the first element, but may need
to be added in both places (TBD).

The goal of this change is to set us up for changing the transformations
to run on both sets of declarations at each step, and then once we hit
SMT encoding we can easily identify which declarations should go in each
of the two SMT environments.
Next steps are hence to change how the transformations are run and then
provide tooling so that we can perform the two-part SMT encoding.