Introduce partition_decls record
Introduce a new record type that separates the different declarations
for a partitioned SRP, to allow us to then distinguish which pieces
should be used for which part of SMT encoding.

Add an initial function to lift transformers over partition_decls, but
it needs additional work to handle questions about map_back
transformers, which need to use solution information.