Flatten ranking in partitioned_decls, add lifting
Change partition_decls to directly represent the lower-ranked hypotheses
and the upper-ranked hypotheses as two lists of declarations, instead of
using a map. This eliminates the complications of using a map and
retaining the rank information, since we only need to know which
hypotheses are lower relative to the given partition.
Additionally, change to storing the symbolic declarations with the rest
of the network, since they are used in all cases and can play a role in
computing the map_back functions.

Update the lifting functions for partitions so that they now work with
this new setup and introduce them throughout the network. Use some
default for simulation and compilation modes, and leave the SMT mode
unfinished. Next steps will involve lifting the slicing behaviors and
finally using the partition_decls properly when encoding the network in
SMT.