Update SMT solve commands to use partitioned_decls
Update the solve functions for SMT to take partitioned_decls instead of
a declaration list. Looks a bit ugly right now since each method now
immediately tests the type of the declarations and fails if it's the
wrong one, but avoids rejiggering the types for a moment.
Also introduces a dependency on kirigami in Nv.smt.

Fixing how the solve functionality works for Kirigami is now the last
piece before the Kirigami encoding should be complete again, with
ideally less necessary processing time.