Fix remaining renaming bugs, perform more interp
Fix a bug with RenameForSMT not properly using the network's environment
for the other declarations in partitioned_decls.
Clean up some of the SMT code.
Add a flag '-print-partitions' to print out the partitioned_decls.
Fix (possibly incorrectly) a bug with CleanupTuples' map back
transformer. No idea if this does the right thing but it seems to work.

Try to perform partial interpretation more aggressively in
TransformDecls. It looks like most of the slowness we're now observing
in larger fattree examples is due to the spine nodes which are now
merging in many other nodes' data. It's not immediately clear if there
is an obvious fix for this, but it could be that we need to revert back
to adding input nodes if there isn't. More partial interpretation could
still be useful if we can do it intelligently.