Remove transcomp, add rtrans/ltrans solve fields
Remove uses of transcomp within NV, replace with two new fields for
solutions, ltrans and rtrans (left and right transfer, respectively).
Users can now specify how they want to decompose the transfer function
using these fields.
With neither specified, the default is to perform the original transfer
function on the output edge, and the identity on the input edge.
With only ltrans or only rtrans given, the other transfer function is
the identity (we assume this usage corresponds with the user
instantiating the field with the trans function).
With both given, we assume the user has a decomposition of the transfer
function they want to use. We make no assumptions that this
decomposition is sound.

An alternative way to provide this feature would be to change the code
so that we have two transfer functions at all times, which are always
composed along non-cut edges and left uncomposed on cut edges. A
possible performance downside of this approach could be if two partial
transfers lead to a larger encoding than one single function.