Drop "assert true" from Kirigami encoding
Remove redundant "assert true" declarations from the Kirigami encoding.
This should shrink down the size of the encoding slightly, which may
improve some solve times marginally.
It looks like larger improvements can be made in the partial
interpreter: statements such as (match foo with | foo -> bar | baz ->
quux) can be easily simplified but are not currently handled. Doing this
should hopefully lead to more significant improvements when dealing with
some of the larger statements we have for the fattree examples (both for
the global network and for the partitioned version!).