Further improve partial interpretation, benchmarks
Make further updates to partial interpreter so that it should now be
able to get rid of unnecessary match statements when using
InterpPartialFull.
Modify the part_nv benchmark generator to copy the program rather than
use an include, which accidentally introduced two solution variables to
encode rather than one.
This has made it possible to complete sp4-part3 in relatively little
time and sp8-part3 in over a minute, but this is still fairly slow
compared to our desired speeds.
Performance is still impacted by map unrolling and some degree of
program size blowup that needs to be dealt with.