Separate hypotheses from base, fix encode_decomp
Finish code (prior to testing) for encoding decomposed tranfer
behaviour. Also add a function to look up the relevant hypotheses for
the input~base edge. Handling predicates is still happening in NV, so
maybe this is all we need now (aside from slight modifications for
transformations) before we can start trying to do encoding with this
approach.
The next step would then be moving partitioning down into the
run_smt_classic function and passing the partitioned_srps to the
solveKirigami function for it to use during encoding.

Other changes include splitting hypotheses out to their own part of the
declaration_groups, which may not be necessary. The goal of this change
was originally to improve our ability to identify which symbolics are
hypotheses and which are not, but naming conventions may be sufficient
to handle this question. It leads to some possible issues if the
generated functions created by transformations have issues, but if we
can successfully handle post-transformation behaviour before encoding,
this is a non-issue.

Other future work also involves changing from using
TransformDecls.outputs_assert to encoding the predicate and using it in
SMT, to avoid having to look up the suitable solve variable.
On a related note, we also need to drop assertions referencing solve
variables that belong to cut nodes once we move partitioning after map
unrolling.