Add initial conjunct pruning code
Add some code to transform assertions by pruning conjuncts in
foldNodes-style assertions that refer to cut nodes.
This code only appears to work for networks without nested conjuncts in
each of the individual assertions over nodes.
If we have a nested foldNodes, this causes the rubric not to align with
the conjuncts properly, particularly if any of the attribute elements
are used as a group (e.g. options become 2-tuples, but then are used
together in a single conjunct, as opposed to separately, which can lead
to rubric misalignment).
An alternative strategy is needed where we only loop over the outermost
foldNodes.