Start using kirigami's solve encoding
Add in code to start using Kirigami's solve encoding. We still aren't
handling output guarantees, but we can for now check that inputs are
working and the safety properties are being modified appropriately.

While the former looks like it may be working, the latter still needs
fixing: as nodes are remapped after partitioning, the (sol-proj)
structure will also have a different labelling: if 1n remaps to 0n, then
we are currently cutting references to 0n (as intended) but are left
with references to 1n when we want to remap those to references to 0n.