Move partitions + hyps earlier, fix some bugs
Move partition creation and hypotheses before map unrolling, so that
they are well-defined for the transformations' map back behaviour. This
doesn't fix all our map back problems, however: solutions that have been
cut lead to issues with map back.
An approach courtesy of Devon is to use a mask transformer and dummy
values to essentially give a bunch of values for Z3 to assign based on
nothing and which don't get used, which can then be returned for the map
back transformers so that they can provide a value for those nodes and
the map back functions still work.
This will be the next thing to try getting done.