Add checking hyps/guars in encode_kirigami
Update Partition code so that we use the new partitioned_srp information
later on (it was being lost before).
Add code to produce a set of SMT expressions for hypotheses and
guarantees that are now actually checked with assertions and added to
the Z3 query.
This appears to work for the minimal correct examples partition-simple
and partition-long, but issues still exist with partition-map and the
fattree benchmarks.

We need to check as well that assertions failing on the guarantees are
correctly reported to the user, and note that we can only accommodate
one partitioned_srp right now despite it possibly having multiple
predicate encodings, depending on the solution we're looking at.
Perhaps, rather than calling update_preds in transform_declarations, we
should call it immediately before encoding?