Fix small encoding bugs with constant ordering
Fix a small bug in ordering with encoding predicates, and move back
towards encoding later as opposed to during solve, as this leads to
assertions occurring and manipulating both scopes rather than the
correct ones (ranked vs. safety).
Also correct a mistake in the interface for partition-map.
There are still issues verifying partition examples for now, as
hypotheses are encoded too early, and guarantee failures are not being
reported, but otherwise it looks promising!