Change interface to use bare fn, fix hyp addition
Change interface to use a bare function type. This is necessary,
unfortunately, because NV will otherwise try to unbox the option
functions and discover it can't come up with a default function for the
None case. This problem exists in standard NV as well, along with most
functional languages: functions don't have a clearly-defined default
value.
Because of this change, some rewrites have to be made for existing
examples. The good news about this is that we actually only care about
the cut edges when looking at an interface, since Partition.update_preds
is only called with the cut edges.
This may also indicate that it's worth considering a different syntax
for interfaces: possibly something more like a map structure?
e.g.
{
  0~1: (fun a -> true)
  1~0: foo
}

The other important change in this commit is that hypotheses are now not
written to the SMT environment until the appropriate time. We are still
not out of the woods until we can get proper counterexamples for
guarantees, but this change appears to work as intended.