Rewrite benchmarks to use bare fns with sed
Call sed on the benchmarks to rewrite uses of Some with functions.
This actually breaks all the benchmarks, since they all end with a
"| _ -> None" case that probably should instead be fixed by hand (since
it's likely to occur in non-interface positions).
We're probably better off deleting and regenerating the relevant
partition examples.