Add symbolics before record unrolling, remap ops
Add hypotheses before record unrolling to improve presentation of
counterexamples. Note that a counterexample reports all hypotheses, not
just the ones that were relevant to this partition.

Also improve the code for remapping expressions containing operations.
We now have a simple procedure for remapping expressions which refer to
cut nodes of the form (n1 = n2), (n1 <n n2) or (n1 <=n n2). Remapping is
only able to simplify these expressions if at least one of the two sides
is a node value: in such cases, if that node has been cut, we can
simplify the entire expression to false.