Change solveKirigami to make 2 get-sat calls
Fix an issue where we weren't actually checking if the guarantees were
satisfiable for kirigami SMT. Implement a variation on Smt.solve which
splits the encoding into two parts and checks satisfiability in between
and after.
Also try to fix some of the confusing responses we get now when a
counterexample is reported: there's a lot of text, and it's now all
totally clear which partition it's related to. A future improvement
would be to add a function for solution remapping to get back the old
node names.