Change Solution to allow remapping of nodes
Solution's nodes field is now an int list: this allows Kirigami to
provide a remapping of the old nodes in the network, so that
counterexamples will now tell the user the original node which had an
issue (rather than the new node).
N.B. This doesn't address issues where nodes' solutions could refer to
new node names, however! A solution could contain a node as a term where
though the solution is printed with the correct global node name, the
term inside is still the remapped node. Solving this kind of issue
requires a deeper dive into handling remapping expressions during
partitioning.