Try to fix sp16-part.nv benchmark
The sp16_control turns out to have some odd discrepancies in terms of
how the file is constructed, so we can't substitute it in directly, as
sp16-malformed does conform to the expected network style, just has
incorrect labels. To try and correct this, this commit substitutes in
the partitioning for the correctly-formed network into the malformed
one.