Modify partition-tags to remove LP
Another crack at pathological examples.
We don't actually need LP to obtain a property that no hypothesis allows
us to verify.
Here we add a new node and end up in a situation where we can get a
spurious route from the other SRP sent to node 1, which means that node
1's solution can't satisfy its guarantee to node 2.
Because of this, we cannot guarantee that node 1 sends a route to node
2, which means we cannot verify the property.