Add ignore statements for use with Kirigami
Add EIgnore expressions, which are used for Kirigami's predicates.
An "ignore e" expression in a predicate is transformed to "true" in
output predicates and "e = default" where default is the default
expression for e's type in input predicates.

Also sprinkle some utilities around, like option to_string and boolean
expressions with ebool.