Allow multiple input and output predicates
Allow multiple predicates to be set for the inputs and outputs. They
will all be included in the SMT query for each solve.
Note that this leads to issues with multiple solves: solves after the
first can add predicates, but these predicates may not be related to the
solutions computable in the other solves.
Fixing this is still to do.

The main reasoning for making this change is to allow for global
invariants to be added to the network, where we check invariants on both
the inputs and outputs.