Add initial global assertion support
Allow a global function to be included as part of a solution, which
enables us to check properties of all nodes (base and auxiliary) in a
partitioned network.
Encoding global assertions is still to do, and may involve using
predicates or alternatively directly doing so in the encoding. This
commit adds some code to TransformDecls to add the predicates to the
partitioned_srp directly.