Implement global assertion encoding
Implement an encoding for global assertions that appears to work.
Reporting counterexamples is a little ugly and unclear right now, but we
nonetheless appear to be able to do so.

The basic approach involves encoding predicates for the global assertion
on every node such that the input behaviour is shared across SMT scopes,
but the global check and the ranked check each encapsulate their
respective assertions.
The global check is tested on both the initial state of each node and
its final state, to ensure that hypotheses can't allow overly-strong
predicates to hold: this is sufficiently strict while still allowing us
to reason about the cases we care about, such as set LP or MED values.

A future change that would be worthwhile at this stage is to put all
kirigami behaviour in its own type that is a field of solve. This would
reduce some of the clutter in Syntax and other lang modules, and help
simplify future changes.
Another worthwhile improvement would be to reporting models, as we don't
distinguish which check failed to the user at present. This makes it
somewhat unclear determining what the cause of an error is.