Change back to inductiveness checks
Change back to using a simple inductiveness check without any ranking.
This change is made under the expectation that we will require users to
only provide networks which satisfy our ordering and monotonicity
properties. A future improvement would be to check these properties as
part of the SMT query.

Make a small modification to partition-tags.nv to allow its assertion to
pass under the new checks.