Change partition to take a 32-bit int
The partition type has been expecting an 8-bit integer, but this wasn't
done for any practical purpose, simply because 8 bits seemed like a
reasonably large number of partitions. While it may not lead to any
practical improvements in the tool to allow integers larger than 8
bits (given the potential cost of verifying a network with that many
partitions), the restriction was unnecessary in the first place and
imposed an additional requirement on how many characters we needed in
each file. This change saves some time and simplifies things.