Change back to interface predicates
This change reverts back to requiring users to provide predicates for
interfaces instead of bare values.
For the single solution check, we can simply specify that we
additionally require that the user's predicates be of the form "x = k".

Also, clarify the warning message when transform_assert doesn't get an
expected assert statement.
It may be worth instead just having the code fail here so that the user
is forced to go fix things.