Add a toggle for ranked/unranked kirigami checks
Add a new commandline argument to call for the old kirigami ranked check
using `-ranked`.
The default is now to simply run everything as a single-scoped SMT
query. This assumes that the given NV file has a unique solution, an
idempotent merge and the predicates are exact values (per latest version
of the paper proofs).
It may be useful at some point to add checks for these things to
determine if the given file needs to be tested using -ranked instead.

Also, remove the deprecated globals elements from the remaining smt code.