Update printing of maps
Change `Printing.map_to_string` to use ';' to separate key-value pairs
instead of '\n', and to also print the default value for the given map.
This completes the "writing" portion of issue #64.
The next step is to fix the "reading" portion, allowing maps printed by
`map_to_string` to also be considered valid NV syntax.