Start nvgen executable to replace gen_part_nv
gen_part_nv is kind of brittle since it's just in Python with regular
expressions to parse the NV file. A better longterm approach now that we
want to potentially generate other kinds of networks automatically is to
integrate something directly with NV and use its libraries. The nvgen
program would be designed to allow us to automatically generate
partitions of networks using simulation and either a schema (like the
fattree schemata) or output from hMETIS or some other mapping.
It would also have tools to extend existing benchmarks with no-transit
properties or hijacking properties.